refactor(HR): Chuyển CandidatesModal.jsx vào /modals
